Extract address mapping to a separate file
